+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## NEXT
+
+- Added the `cai_causal_graph.causal_graph.CausalGraph.get_nondirected_edges` method, which returns all edges that are
+  not explicitly directed, i.e. not of type `EdgeType.DIRECTED_EDGE`.
+
 ## 0.5.7
 
 - Added support for `numpy < 3.0.0` for Python versions `>= 3.10`.